delete like record function is now in a seperate function to make it clearer that two functions are sharing the same underpinning
